100% fixed now and modularized code
