+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "apollo-server": "^2.11.0",
     "dotenv": "^8.2.0",
+    "graphql-import": "^1.0.1",
     "prisma-client-lib": "^1.34.10"
   },
   "devDependencies": {
